FGPA Bug addressed (I think. See description)
It would appear that a major bug that was screwing up my signal generation was that the syntax vector1[reg+:const] <= vector2; does something different in simulation than hardware. In simulation, if reg+const spills past the width of vector1, then those bits are just dropped, as I would expect. But in the hardware, it appears to wrap around and overwrite values below reg. Idk why this would happen and I especially don't know why it would be different in hardware than simulation, but hopefully this fixes the problem